requirements.txt
Django==1.8
bleach==1.4.1
django-html-sanitizer==0.1.4
django-pipeline==1.4.7
futures==2.2.0
html5lib==0.999
psycopg2==2.6
six==1.9.0
uWSGI==2.0.10
